# docs/design/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NotificationPublic": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "fit_level": {
            "enum": [
              "high",
              "medium"
            ],
            "examples": [
              "high"
            ],
            "title": "Fit Level",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_read": {
            "default": false,
            "title": "Is Read",
            "type": "boolean"
          },
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "opportunity_id": {
            "format": "uuid",
            "title": "Opportunity Id",
            "type": "string"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Read At"
          },
          "title": {
            "examples": [
              "Opportunity matches your unit"
            ],
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "const": "opportunity_match",
            "default": "opportunity_match",
            "title": "Type",
            "type": "string"
          },
          "unit_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Id"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "created_at",
          "id",
          "fit_level",
          "opportunity_id",
          "title",
          "message"
        ],
        "title": "NotificationPublic",
        "type": "object"
      },
...
      "OpportunityCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "client": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpportunityParty"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "extracted": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpportunityExtracted"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured fields if already extracted client-side; server may re-evaluate."
          },
          "source": {
            "default": "chat",
            "enum": [
              "chat",
              "hubspot",
              "manual"
            ],
            "title": "Source",
            "type": "string"
          },
          "title": {
            "examples": [
              "D365 implementation for retail client"
            ],
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "description"
        ],
        "title": "OpportunityCreateRequest",
        "type": "object"
      },
...
      "OpportunityPublic": {
        "additionalProperties": false,
        "properties": {
          "client": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpportunityParty"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "extracted": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpportunityExtracted"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_official": {
            "default": false,
            "title": "Is Official",
            "type": "boolean"
          },
          "pushed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pushed At"
          },
          "source_ref": {
            "$ref": "#/components/schemas/SourceRef"
          },
          "status": {
            "default": "draft",
            "enum": [
              "draft",
              "open",
              "won",
              "lost",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "created_at",
          "id",
          "title",
          "description",
          "source_ref"
        ],
        "title": "OpportunityPublic",
        "type": "object"
      },
      "OpportunityPushCrmRequest": {
        "additionalProperties": false,
        "properties": {
          "confirm": {
            "default": false,
            "description": "Must be true to execute CRM write.",
            "examples": [
              true
            ],
            "title": "Confirm",
            "type": "boolean"
          }
        },
        "title": "OpportunityPushCrmRequest",
        "type": "object"
      },
      "OpportunityPushCrmResult": {
        "additionalProperties": false,
        "properties": {
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "HubSpot deal id when success.",
            "title": "External Id"
          },
          "message": {
            "examples": [
              "Pushed to HubSpot"
            ],
            "title": "Message",
            "type": "string"
          },
          "source": {
            "default": "hubspot",
            "enum": [
              "chat",
              "hubspot",
              "manual"
            ],
            "title": "Source",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message"
        ],
        "title": "OpportunityPushCrmResult",
        "type": "object"
      },
      "OpportunityUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "client": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpportunityParty"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 10000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "extracted": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OpportunityExtracted"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "draft",
                  "open",
                  "won",
                  "lost",
                  "archived"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "OpportunityUpdateRequest",
        "type": "object"
      },
      "Paginated_NotificationPublic_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationPublic"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Total number of matching items.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Paginated[NotificationPublic]",
        "type": "object"
      },
      "Paginated_OpportunityPublic_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OpportunityPublic"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Total number of matching items.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Paginated[OpportunityPublic]",
        "type": "object"
      },
...
      "UnitPublic": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "$ref": "#/components/schemas/UnitCapabilities"
          },
          "code": {
            "examples": [
              "D365",
              "AI",
              "JAPAN"
            ],
            "maxLength": 50,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "contact": {
            "$ref": "#/components/schemas/UnitContact"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "examples": [
              "Rikkei D365 Division"
            ],
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "status": {
            "default": "active",
            "enum": [
              "active",
              "inactive"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "created_at",
          "id",
          "code",
          "name",
          "contact",
          "capabilities"
        ],
        "title": "UnitPublic",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Bear In Mind API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/notifications": {
      "get": {
        "description": "US2: FE polls this endpoint; later can evolve to push/stream.",
        "operationId": "list_notifications_api_v1_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "unread_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unread Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_NotificationPublic_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List notifications (polling)",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "operationId": "mark_read_api_v1_notifications__notification_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notification Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark a notification as read",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/v1/opportunities": {
      "get": {
        "description": "US6: returns a consolidated list; v1 can start with local DB only.",
        "operationId": "list_opportunities_api_v1_opportunities_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "unit_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Unit Id"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_OpportunityPublic_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List opportunities (multi-source later)",
        "tags": [
          "opportunities"
        ]
      },
      "post": {
        "description": "US5: persist an opportunity captured from chat or manual entry.",
        "operationId": "create_opportunity_api_v1_opportunities_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpportunityCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create an opportunity (draft/unofficial supported)",
        "tags": [
          "opportunities"
        ]
      }
    },
    "/api/v1/opportunities/{opportunity_id}": {
      "get": {
        "operationId": "get_opportunity_api_v1_opportunities__opportunity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get opportunity by id",
        "tags": [
          "opportunities"
        ]
      },
      "put": {
        "operationId": "update_opportunity_api_v1_opportunities__opportunity_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpportunityUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update an opportunity",
        "tags": [
          "opportunities"
        ]
      }
    },
    "/api/v1/opportunities/{opportunity_id}/push-crm": {
      "put": {
        "description": "US5: requires explicit confirmation via request body.",
        "operationId": "push_opportunity_to_crm_api_v1_opportunities__opportunity_id__push_crm_put",
        "parameters": [
          {
            "in": "path",
            "name": "opportunity_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Opportunity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpportunityPushCrmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpportunityPushCrmResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Push an opportunity to CRM (HubSpot)",
        "tags": [
          "opportunities"
        ]
      }
    },
    "/api/v1/units": {
      "get": {
        "description": "Useful for FE dropdowns and admin screens.",
        "operationId": "list_units_api_v1_units_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UnitPublic"
                  },
                  "title": "Response List Units Api V1 Units Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List units",
        "tags": [
          "units"
        ]
      }
    },
    "/api/v1/units/{unit_id}": {
      "get": {
        "operationId": "get_unit_api_v1_units__unit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "unit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Unit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get unit by id",
        "tags": [
          "units"
        ]
      }
    },
...
  }
}